refactor: move connpass/icalendar/archive into app/providers/
These three modules share one shape (an *EventRequest/*IndexRequest
class that fetches from an external source with cache-backed
get_events/get_groups), and app/ was accumulating enough top-level
files that the grouping was getting lost. Moves them under
app/providers/ and updates the importers (service.py, tests) and the
relative .models imports (now two levels up) accordingly.
